address qa feedback
