+++ apps/web/src/components/Developers/Shared/CtaBanner.tsx
@@ -13,11 +13,15 @@ type CtaBannerProps = {
 export function CtaBanner({ title, description, cta }: CtaBannerProps) {
   return (
     <section className="w-full bg-black">
-      <div className="flex flex-col rounded-2xl bg-dark-palette-backgroundAlternate px-8 py-8 sm:items-center sm:py-16">
-        <Title level={TitleLevel.Title1} as="h2" className="max-w-2xl sm:text-center max-sm:hidden">
+      <div className="flex flex-col items-center rounded-2xl bg-dark-palette-backgroundAlternate px-8 py-16 py-8">
+        <Title level={TitleLevel.Title1} as="h2" className="max-w-2xl max-sm:hidden sm:text-center">
           {title}
         </Title>
-        <Title level={TitleLevel.Title3} as="h2" className="max-w-2xl sm:text-center sm:hidden font-bold">
+        <Title
+          level={TitleLevel.Title3}
+          as="h2"
+          className="max-w-2xl text-center font-bold sm:hidden"
+        >
           {title}
         </Title>
         {description && (
