+++ apps/web/src/components/Developers/SmartWallet/onboarding.png
